docs: refactor Player example
