+++ packages/docs/components/ColorDemo/Name.tsx
@@ -6,12 +6,7 @@ import {
   useCurrentFrame,
   useVideoConfig,
 } from "remotion";
-
-const container: React.CSSProperties = {
-  fontFamily:
-    "--apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif",
-  fontWeight: "bold",
-};
+import { defaultStyles } from './styles';
 
 export const Name: React.FC<{
   name: string;
@@ -28,25 +23,23 @@ export const Name: React.FC<{
   return (
     <AbsoluteFill
       style={{
-        ...container,
-        display: "flex",
-        justifyContent: "center",
-        alignItems: "center",
-        flex: 1,
-        fontSize: 100,
+        ...defaultStyles,
         backgroundColor: "white",
         color: "black",
       }}
     >
-      <div>
+      <div
+        style={{
+          transform: `translateY(${interpolate(
+            progress,
+            [0, 1],
+            [1000, 0]
+          )}px)`,
+        }}
+      >
         <span
           style={{
             display: "block",
-            transform: `translateY(${interpolate(
-              progress,
-              [0, 1],
-              [1000, 0]
-            )}px)`,
             lineHeight: 1.1,
           }}
         >
@@ -55,11 +48,6 @@ export const Name: React.FC<{
         <span
           style={{
             display: "block",
-            transform: `translateY(${interpolate(
-              progress,
-              [0, 1],
-              [1000, 0]
-            )}px)`,
             lineHeight: 1.1,
           }}
         >
