+++ packages/docs/components/ColorDemo/index.tsx
@@ -1,13 +1,13 @@
-import {parseToRgb} from 'polished';
+import { parseToRgb } from 'polished';
 import React from 'react';
-import {AbsoluteFill, Sequence} from 'remotion';
-import {Big} from './Big';
-import {Colors} from './Colors';
-import {Name} from './Name';
-import {Palette} from './Palette';
-import {Shade} from './Shade';
+import { AbsoluteFill, Sequence } from 'remotion';
+import { Big } from './Big';
+import { Colors } from './RGBColors';
+import { Name } from './Name';
+import { Palette } from './Palette';
+import { Shade } from './ColorName';
 
-export const Main: React.FC<{
+export const ColorDemo: React.FC<{
 	color: string;
 	name: string;
 }> = ({color, name}) => {
