+++ packages/docs/components/ColorDemo/styles.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+export const defaultStyles: React.CSSProperties = {
+	display: "flex",
+	flex: 1,
+	justifyContent: "center",
+	alignItems: "center",
+	fontFamily:
+	"--apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif",
+	fontWeight: "bold",
+	fontSize: 100,
+	lineHeight: 1.1,
+};
